requirements.txt
influxdb-client
smllib
pyserial-asyncio
docopt